refactor : 레이싱 결과 클래스 추가에 따른 기능 분리 feat: 레이싱 결과 클래스 추가
* 자동차 저장소 클래스 기능 분리
* List<Car>로 반환하던 로직을 RaceResult를 반환하면서 의도가 명확해지고
기능이 축약되었음.
* 일급 컬렉션에 대해 몸소 느끼는 중
